Finished!
I know there are some problems still, but I finished the basic form of Tic-Tac-Toe.
So at least, I'm glad to say this is a good news!

In this released ver1.0 I try to use pointer (*)which is not used in MineAngel before. To be honest, that's the most difficult part of C and also of my program.

The same time, the logical part is not as simple as I thought. espacially I made it changable, which means this program can not only played on 3x3 metrix but smaller or larger. That's up to you. And may I will add this selection into in the late version, perhaps?

Anyway, this is exactlly the first game I made by myself, without any reference (OK maybe GPT.....XD), but the structure, the idea, 98% of this program is made by myself. I promise.

Hope I can make my own game asap. It's aleady on the draft! Hope we can meet her before I gruaduate!

Peace and love!